refactor: colocate stories with their components
